fix: auto grant matched oauth credentials
