+++ services/console/app/models/static_secret.rb
@@ -48,6 +48,10 @@ class StaticSecret < ApplicationRecord
   # proxy resolves at sync; this association is the console-level link.
   belongs_to :broker_credential, optional: true
 
+  after_commit :auto_grant_wrapped_oauth_credential,
+               on: %i[create update],
+               if: :broker_credential_id?
+
   # Maps to a single entry in the iron-proxy `secrets` transform array. The
   # caller is responsible for skipping secrets without a source.
   def to_proxy_secret
@@ -90,6 +94,10 @@ class StaticSecret < ApplicationRecord
   validate :replace_config_matches_schema
 
   private
+
+  def auto_grant_wrapped_oauth_credential
+    PrincipalCredentialReconciliation.new.apply_for_credential(broker_credential)
+  end
 
   def labels_is_a_hash
     errors.add(:labels, "must be a hash") unless labels.is_a?(Hash)
