--- services/console/app/views/console/principal_reconciliations/index.html.erb
@@ -1,123 +1,0 @@
-<% content_for :title, "Principal Reconciliation · Centaur Console" %>
-
-<div class="mb-8">
-  <h1 class="page-title">Principal Reconciliation</h1>
-  <p class="page-subtitle">
-    <%= pluralize(@entries.size, "user principal") %> with matched Slack or Google OAuth credentials.
-  </p>
-</div>
-
-<div class="mb-6 grid gap-4 sm:grid-cols-3">
-  <div class="rounded border border-ink-600 bg-ink-850/60 px-4 py-3">
-    <div class="text-xs uppercase tracking-wider text-zinc-500">Slack credentials</div>
-    <div class="mt-1 text-2xl text-zinc-100"><%= @entries.sum { |entry| entry.slack_credentials.size } %></div>
-  </div>
-  <div class="rounded border border-ink-600 bg-ink-850/60 px-4 py-3">
-    <div class="text-xs uppercase tracking-wider text-zinc-500">Google credentials</div>
-    <div class="mt-1 text-2xl text-zinc-100"><%= @entries.sum { |entry| entry.google_credentials.size } %></div>
-  </div>
-  <div class="rounded border border-ink-600 bg-ink-850/60 px-4 py-3">
-    <div class="text-xs uppercase tracking-wider text-zinc-500">Grantable missing</div>
-    <div class="mt-1 text-2xl text-zinc-100"><%= @entries.sum { |entry| entry.actionable_credentials.size } %></div>
-  </div>
-</div>
-
-<div class="console-panel">
-  <table class="console-table">
-    <thead class="console-table-head">
-      <tr>
-        <th class="console-th">Principal</th>
-        <th class="console-th">Slack OAuth</th>
-        <th class="console-th">Google OAuth</th>
-        <th class="console-th">Grant State</th>
-        <th class="console-th text-right"></th>
-      </tr>
-    </thead>
-    <tbody class="divide-y divide-ink-700">
-      <% if @entries.empty? %>
-        <tr><td colspan="5" class="px-5 py-10 text-center text-zinc-600">No OAuth credentials matched user principals yet.</td></tr>
-      <% end %>
-
-      <% @entries.each do |entry| %>
-        <tr class="console-row-hover align-top">
-          <td class="console-td">
-            <%= link_to entry.principal.name.presence || entry.principal.foreign_id.presence || entry.principal.oid,
-                        console_principal_path(entry.principal.oid),
-                        class: "console-link" %>
-            <%= id_meta_line(entry.principal.namespace, oid: entry.principal.oid) %>
-            <% if entry.principal.foreign_id.present? %>
-              <div class="mt-1 text-xs text-zinc-500"><%= entry.principal.foreign_id %></div>
-            <% end %>
-          </td>
-
-          <td class="console-td">
-            <% if entry.slack_credentials.any? %>
-              <div class="space-y-2">
-                <% entry.slack_credentials.each do |credential| %>
-                  <div>
-                    <%= link_to credential.name.presence || credential.provider_subject || credential.oid,
-                                console_credential_path(credential.oid),
-                                class: "console-link" %>
-                    <div class="mt-0.5 text-xs text-zinc-500">
-                      <%= credential.provider_email.presence || "no email" %>
-                    </div>
-                  </div>
-                <% end %>
-              </div>
-            <% else %>
-              <span class="text-zinc-600">missing</span>
-            <% end %>
-          </td>
-
-          <td class="console-td">
-            <% if entry.google_credentials.any? %>
-              <div class="space-y-2">
-                <% entry.google_credentials.each do |credential| %>
-                  <div>
-                    <%= link_to credential.name.presence || credential.provider_email || credential.oid,
-                                console_credential_path(credential.oid),
-                                class: "console-link" %>
-                    <div class="mt-0.5 text-xs text-zinc-500">
-                      <%= credential.provider_email.presence || "no email" %>
-                    </div>
-                  </div>
-                <% end %>
-              </div>
-            <% else %>
-              <span class="text-zinc-600">missing</span>
-            <% end %>
-          </td>
-
-          <td class="console-td">
-            <div class="flex flex-wrap gap-1">
-              <% entry.credentials.each do |credential| %>
-                <% provider = credential.oauth_app.provider %>
-                <% granted = entry.granted?(credential) %>
-                <% has_wrapper = credential.static_secret.present? %>
-                <span class="inline-flex items-center rounded px-2 py-0.5 text-[11px] ring-1 <%= granted ? "bg-centaur-500/10 text-centaur-300 ring-centaur-500/30" : has_wrapper ? "bg-amber-500/10 text-amber-300 ring-amber-500/30" : "bg-red-500/10 text-red-300 ring-red-500/30" %>"
-                      title="<%= credential.oid %>">
-                  <%= provider %>: <%= granted ? "granted" : has_wrapper ? "ready" : "no wrapper" %>
-                </span>
-              <% end %>
-            </div>
-          </td>
-
-          <td class="console-td text-right">
-            <% actionable = entry.actionable_credentials %>
-            <% if actionable.any? %>
-              <%= form_with url: console_principal_reconciliation_apply_path, method: :post, data: { turbo: false }, class: "inline-block" do %>
-                <%= hidden_field_tag :principal_id, entry.principal.oid %>
-                <% actionable.each do |credential| %>
-                  <%= hidden_field_tag "credential_ids[]", credential.oid %>
-                <% end %>
-                <%= submit_tag "Grant missing", class: "btn-primary" %>
-              <% end %>
-            <% else %>
-              <span class="text-sm text-zinc-600">settled</span>
-            <% end %>
-          </td>
-        </tr>
-      <% end %>
-    </tbody>
-  </table>
-</div>
