+++ services/console/app/views/layouts/console.html.erb
@@ -37,7 +37,6 @@
               <nav class="flex items-center gap-1">
                 <% nav = [
                   ["principals", console_principals_path],
-                  ["reconcile", console_principal_reconciliation_path],
                   ["roles", console_roles_path],
                   ["secrets", console_secrets_path],
                   ["credentials", console_credentials_path],
